pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "fastapi-crud-service"
version = "1.0.0"
description = "Production-ready FastAPI CRUD service with JWT auth and PostgreSQL"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.23,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "bcrypt>=4.0.0,<6.0.0",
    "python-multipart>=0.0.6,<1.0.0",
    "email-validator>=2.1.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "httpx>=0.25.0,<1.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "sqlalchemy-stubs>=0.4",
    "types-python-jose>=3.3.0",
    "coverage>=7.3.0",
    "aiosqlite>=0.19.0,<1.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "T20"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = []
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["jose.*", "bcrypt.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
concurrency = ["greenlet"]
omit = [
    "tests/*",
    "alembic/*",
    "app/database.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "def __repr__",
]
